moved particle objects into a separate folder, small fixes
