+++ src/io/input/FileReader.cpp
@@ -6,8 +6,8 @@
  */
 
 #include "FileReader.h"
-#include "simulator/ParticleContainer.h"
-#include "simulator/ParticleGenerator.h"
+#include "simulator/particle/ParticleContainer.h"
+#include "simulator/particle/ParticleGenerator.h"
 
 #include <cstdlib>
 #include <fstream>
