+++ src/io/output/XYZWriter.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "simulator/ParticleContainer.h"
+#include "simulator/particle/ParticleContainer.h"
 #include "FileWriter.h"
 
 #include <fstream>
